// packages/fhir-r5/src/v5.0.0/structuremapparameter/types.ts
import type { Extension, Element } from "../core/types";

/** Generated from FHIR JSON Schema */

/** A Map of relationships between 2 structures that can be used to transform data. */
export interface StructureMapParameter {
  id?: string;

  extension?: Extension[];

  modifierExtension?: Extension[];

  valueId?: string;

  _valueId?: Element;

  valueString?: string;

  _valueString?: Element;

  valueBoolean?: boolean;

  _valueBoolean?: Element;

  valueInteger?: number;

  _valueInteger?: Element;

  valueDecimal?: number;

  _valueDecimal?: Element;

  valueDate?: string;

  _valueDate?: Element;

  valueTime?: string;

  _valueTime?: Element;

  valueDateTime?: string;

  _valueDateTime?: Element;
}